refactor: share chunkText() from @mulmobridge/client
Extract identical chunkText into packages/client/src/text.ts.
Remove 5 local copies in messenger, mattermost, zulip, line, relay.
