+++ packages/relay/package.json
@@ -31,5 +31,8 @@
   "engines": {
     "node": ">=20"
   },
-  "license": "MIT"
+  "license": "MIT",
+  "dependencies": {
+    "@mulmobridge/client": "^0.1.1"
+  }
 }
